Move: Personal details mapped into mocks folder
